Move parallel oidc tests
